+++ CHANGELOG.md
@@ -10,9 +10,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add an `AddressesManagement` component
 - Add a `PaymentButton` component 
 - Add a `PaymentInterface` component to lazy load the right payment component
-  according to the provider used 
+  according to the provider used
 - Create a `StepBreadcrumb` component to display progress within a step process
 - Create a `useStepManager` hook to manage step process
 - Create a React `Icon` component that can optionally take alternative text